requirements.txt
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
click==8.1.3
rich==13.3.5
PyYAML==6.0.1
